+++ src/interfaces/ICustomerActions.java
@@ -2,7 +2,7 @@ package interfaces;
 
 public interface ICustomerActions {
     double checkBalance(String name);
-    void deposit(String name, double amount);
-    boolean withdraw(String name, double amount);
+    void deposit(String name, int amount);
+    boolean withdraw(String name, int amount);
     void printReceipt();
 }
